Extract parsing the grouping to an external file.
